VIC-5616 Exploring DAS: info about environments (#3133)
As part of exploring DAS events, send info on the robot's memory map
whenever the robot is delocalized*. This sends the area of known space,
the area that comprises obstacles, and how long the map was used before
delocalization, so that we can learn about trends of use and operating
environments, like table size and clutter. There are also a handful of
new methods to query the entire navmap instead of just a region.
* If the robot is put down, even though the robot is then considered
delocalized, we skip sending the map info, since that map would only
contain stats from when the robot in the air. Instead, info is only
sent when the robot is initially picked up/bumped/etc.