VIC-5762 Use full popawheelie animation
Currently any popawheelie action is aborted when the robot is on its
back, and the reaction to being on the robot back is played to get
the robot down.
To play the entire popawheelie action in onboarding, this suppresses
any reaction to being on the back from predock to completion.