VIC-5760 Loading animations at runtime is blocking (#3221)
Loading JSON can take a long time, enough for the anim process to be considered inactive and terminated.

This change executes _context->GetDataLoader()->LoadAnimationFile on a separate job thread.
The animation loading themselves are done on worker threads, and hence are thread safe, but the
controlling function is synchronous as it waits for the animations to complete loading, in the case
of a single animation it's effectively a synchronous load.

Also tweaked two asserts that can fire when animations are triggered on-demand via AddAnimaition
instead of the usual pipeline.

Moved the PlayAnimation onto the same thread as the LoadAnimation, which now occurs inline with
all of the loads, i.e. does not require any special guards or handling. Implemented by
introducing a `pending` animation and num loops that are mutex protected and applied during
the AnimationStreamer ::Update; the alternative was to add a mutex to all references to
_streamingAnimation and there are many.