VIC-6082 Add user agent to jdoc requests (#3317)
I forgot to add the OSUserAgent() dial option to the jdocs
connection, so server logs for jdoc requests aren't showing OS
information. To simplify this in the future and reduce the need
for a "blah_options_vicos.go" file in every service, this change
moves the definition of GRPC options used across all services to
a file in the util package (and uses it in in jdocs).