VIC-6279 Fix voice request never closing if no audio received (#3340)
If a voice request never sends up any audio, it never starts the
"wait for response" routine. This change adds a routine that
listens for a stream's context to complete, and if the completion
was caused by a timeout and not the closure of the stream, returns
an error to the main process.