VIC-3701 BlackJack prompt hit or stand thrice
This is essentially a revert of the last commit to this jira tag.
Didn't notice the age of the ticket... or check the logic for
validity. Just thought, it was wrong at two, should be three. But the
setting is number of REprompts, so the ticket was invalid a while ago.
The number of thy hit-stand promptings shall be three.