VIC-6346: Head periodically engages during BehaviorOnboardingLookAtPhone (#3351)
This happens if the head is miscalibrated because a head keyframe in anim_lookatphone_loop_01 periodically tries to drive the head up. If uncalibrated, the head thinks it can't reach the angle and MotorBurnoutProtection causes the head to disable temporarily because it thinks it's being held (and therefore someone is probably messing with the head).

Best fix seems to be to make the IsBeingHeld() check not care if motors are powered, only if they're moving. That way when the robot is on a table and the head (or lift) is pushing against a hard limit, IsBeingHeld() would be false and MotorBurnoutProtection would trigger calibration.