VIC-6097 Fail CloudSessionRequest immediately if offline
In the case where the robot is offline, sending a CloudSessionRequest to `vic-cloud` can
result in a 30+ sec delay[1]. During this time, the app is connected over BLE and is
showing a connection modal dialog, indicating a connection is in progress. The result for end
users of Chewie (app) is that it appears to take a long to time connect over BLE, since the
UI does not indicate that the robot is connected. Furthermore, in this case, the robot knows
it is not connected to Wi-Fi. Thus, in the case where we are not connected to an AP, we
now skip the call to `vic-cloud` and immediately return an error to the client.

To further avoid problems, the app should also consider implementing a `status` request
after connecting over BLE. This can be used to determine whether the robot is owned or
not. If the robot is not owned, the status result will indicate the connection state.

[1]: 40 sec delay between cloud session auth request by `vic-switchboard` and
response by `vic-cloud`.

```
08-30 00:16:23.510  2014  2014 I vic-switchboard: Received cloud session authorization request.
08-30 00:16:23.525  2014  2014 I vic-switchboard: Received message from token_server: 4
08-30 00:16:23.525  2014  2014 I vic-switchboard: CloudRequest JWT Response Handler
08-30 00:16:37.124  2014  2014 I vic-switchboard: WifiWatcher: detected no wifi. Scanning for networks...
08-30 00:16:37.168  2014  2014 I vic-switchboard: Not in access point mode or could not disable AccessPoint mode
08-30 00:16:52.140  2014  2014 I vic-switchboard: WifiWatcher: detected no wifi. Scanning for networks...
08-30 00:16:52.185  2014  2014 I vic-switchboard: Not in access point mode or could not disable AccessPoint mode
08-30 00:17:03.578  2019  2053 I vic-cloud: Token queue error: rpc error: code = Unavailable desc = all SubConns are in TransientFailure, latest connection error: connection error: desc = "transport: Error while dialing dial tcp: i/o timeout"
08-30 00:17:03.581  2019  2056 I vic-cloud: Error handling token request: rpc error: code = Unavailable desc = all SubConns are in TransientFailure, latest connection error: connection error: desc = "transport: Error while dialing dial tcp: i/o timeout"
08-30 00:17:03.672  2014  2014 I vic-switchboard: Received message from token_server: 6
08-30 00:17:03.672  2014  2014 E vic-switchboard: CloudAuth - vic-cloud could not connect to server.
```

(cherry picked from commit 703c7978b065b586b85eb12d0ad3ad03a4e89ff0)
Signed-off-by: Naveen Akkapeddi <naveen.akkapeddi@anki.com>