CLAI-225 Add connection testing to web server (#3365)
This change adds a home page to the cloud web server (which
previously only existed to serve the token authorization endpoint
for tokprovision), as well as a new page that reports results from
constantly running connection attempts in the background.

Every 5 seconds, and only in dev builds, the robot will attempt to
open a connection to the OTA server. It will log whether the
attempt succeeded, along with the number of milliseconds it took
or an error if one occurred. Connecting to <IP>:8890/connect.html
will then show the results of these tests.

Example output (after I intentionally disconnected from wifi in
the middle for a bit):

Results for last 523 connections
17:36:01-18:14:01: 457 successes (min time 23 ms, max 324, avg 42)
18:14:06: ERROR: dial tcp: lookup ota-cdn.anki.com on [::1]:53: server misbehaving
18:14:11-18:14:56: 10 successes (min time 25 ms, max 82, avg 41)
18:15:01: ERROR: dial tcp: lookup ota-cdn.anki.com on [::1]:53: server misbehaving
18:15:06-18:18:11: 38 successes (min time 25 ms, max 977, avg 64)
18:18:16: ERROR: dial tcp: lookup ota-cdn.anki.com on [::1]:53: server misbehaving
18:18:21: ERROR: dial tcp: lookup ota-cdn.anki.com on [::1]:53: server misbehaving
18:18:26-18:18:41: 4 successes (min time 26 ms, max 74, avg 44)
18:18:46: ERROR: dial tcp: lookup ota-cdn.anki.com on [::1]:53: server misbehaving
18:18:51: ERROR: dial tcp: lookup ota-cdn.anki.com on [::1]:53: server misbehaving
18:18:56-18:19:31: 8 successes (min time 24 ms, max 53, avg 35)

* Make a HTTP HEAD request instead of opening/closing TCP