VIC-3932 Disable unnecessary gRPC and REST endpoints for 1.0 (#3382)
This needs to be in the 1.0 branch and any releases until the Vector Python release is done with the holiday release. This is committed to master only for the purposes of getting into those releases. After the 1.0 branch, this will be reverted on master.

(cherry picked from commit 272c9f04d2a103b45e009e780b7c75ea53722ff8)
Signed-off-by: David Woods <dwoods@anki.com>