VIC-6441: Disable punctuation pause to avoid trailing silence (#3441)
@jmrivas discovered that Acapela TTS adds a punctuation pause at the end of each utterance, even if text does not contain any punctuation marks.
This causes unwanted trailing silence.

We want punctuation marks to add silence WITHIN an utterance but not at the end.  I opened a ticket with Acapela to find out how we can do this.

Workaround is to disable code that sets Acapela PAUSE1SILENCE.
We still get a minimal (50ms) trailing silence but we don't get the full pause.