VIC-6578 Fix drive off charger timeout (#3498)
If drive off charger animations fail, a DriveStraightAction
will repeatedly try to get the robot off the charger, up to some
max number of attempts. However, the timeout was 30 seconds, so it
would take a long time.
This reduces the DriveStraightAction timeout in this case to the
expected time to drive off the charger.