VIC-6554: Remove cv::warpAffine for faster eye rendering (#3491)
For better face rendering performance, this completely removes the TransformFace method and its call to the expensive `cv::warpAffine` method. Instead, we now compose the full-face transformation parameters with those for each eye, and directly warp the eye and lid polygons directly up front. The slightly trickier bit was to also adjust the covariance matrix defining the Gaussian falloff of the eyes' hotspots as well. 

Because this gives us back a little CPU and has a unit test to help verify that the rendering differences are imperceptible, I think it's a candidate for 1.0.

* Main change is what's described above
  - There's a minor API change to use 3x3 matrices for the warps instead of 2x3, just to make composition easier.
  - Needed a small change to HSV to RGB565 conversion code to use signed ints 
* Updates the unit tests to use a better image-difference metric and replaces the unit test images with newly-generated ones.
* Note that the differences between the old and new unit test images are nearly imperceptible. I would actually argue the new method is more "correct" in that the scaling and interpolation previously done by cv::warpAffine was actually introducing blurring/sharpening artifacts which the new method avoids.
* I used CpuProfiler to confirm that DrawFace was clocking at 4-5.5ms before and 2.5-3.5ms after.