VIC-4547 Fix go fault-code reporting (#3515)
Golang fmt strings do not accept `%u` as a valid format for unsigned ints.
Using it seems to work sometimes, but other times it appear to return an
escaped value of the literal input, for example:

```
%!u(uint16t=850)
```

Our fault-code-handler was removing, non-digit chars and displaying `16850`
as the code, which should have been `850`. This change uses `%d`, which is
the correct fmt specifier for all (u)int types.