VIC-6698 Fix behavior after entering CC screen (#3541)
If you "complete" onboarding but are still in the onboarding stack,
then entering and leaving the CC screen puts you back at the start of
onboarding because the boot behavior is still onboarding.
This changes the return value of GetBootBehavior() whenever the
onboarding stage is changed, but keeps the behavior stack the same
until it is next reset.

(cherry picked from commit 42cdb83449ed0ae0e3a3f117433e37c401b2c369)
Signed-off-by: David Woods <dwoods@anki.com>