VIC-1951 VIC-6721 Safe readEsn + group dir access (#3516)
* VIC-1951 Use `emr-cat` instead of reading raw disk

When running as a non-root user, `vic-cloud` does not have permission
to read from raw block partitions.  Instead, it should use the `emr-cat`
utility, which was created for the purpose of accessing `emr` data in
a consistent way for every program. This utility can have its sticky
bit set to run with appropriate permissions to access the block partition
and return data.

* VIC-1951 Create dirs with group rwx privs

All vic-* share a common location in /data (/data/data/com.anki.victor)
where persistent or cache files live. When programs are running as non-root
users, they will still belong to the `anki` group so that they can access
any shared persistent data. This change ensures that the `persistent`,
`cache` and other dirs created under our shared data dir are accessible to
all programs that are members of the `anki` group.

* VIC-1951 Add mode param to FileUtils::CreateDirectory

The default value is S_IRWXU|S_IRWXG. Since this is intended to be used for anki code,
which will be run by users that share a common group, it will grant all users in the
`anki` group full access to any directories that might be created. Code that requires
stricter access can change the param (such as dasmgr). Process that require stricter
default values can set the `umask`, which will be used by the underlying `mkdir` call.