VIC-6786 Increase IPC buffer size (#3568)
To put duct tape on the issue of the app tokens jdoc getting too
large to transmit over IPC after clearing user data 67 times, this
change increases the buffer size from 8k to 32k for 1.0. A better
fix making IPC more resilient for large messages will come later
for 1.1, but is too large to do this late for 1.0.