VIC-6312 [Timer] Error 915 displays on Vector's face when canceling the timer (#3567)
When a timer is cancelled a large number of 1023 byte messages are placed on
the queue between vic-engine and vic-anim at 1 ms intervals which are impossible
for vic-anim to process. The queue is filled and the Local UDP code gets a
`EDEADLK` or `EAGAIN` and returns -1. When `AnimComms::SendPacketToEngine`
recieves the -1 it closes the connection which ultimately yields a 915/917 error.

Rather than fail on `EDEADLK` or `EAGAIN` this change adds a retry and delay
to give the queue a chance to free up some space. The total time for retry is
kept significantly under 5 seconds which is the maximum pause observed before
915/917 errors have occured (see through inspection of Jira tickets and log files).

Testing for the timer case a maximum of 3 retries was observed.

This fix should alleviate 915/917 errors for the same condition. And the AI team
are looking to change how timer cancel behaves in vic-engine.