VIC-6787 Refactor IPC to use fewer buffers (#3569)
To fix the issue of the app tokens jdoc overflowing an 8k buffer,
PR #3568 increases the buffer size to 32k. However, since that
change would apply to ~12 buffers on its own, this change also
takes steps to reduce the number of buffers from 12 to 9.

Currently, every connection (client ond server alike) launches a
reader routine which allocates a buffer for read messages. On top
of that, datagram servers also have a reader routine which they use
to read messages before sending them to whichever of their spawned
clients the message was directed to.

This creates some redundancy, because client connections spawned
from datagram servers have that internal buffer for reading
messages, even though the place they're reading from (the datagram
server) has already utilized its buffer to read the message. That
meant the previous implementation of `packetClient.Read` looked
like this:

```go
func (c *packetClient) Read(buf []byte) (int, error) {
	select {
	case <-c.kill:
		return 0, io.EOF
	case recv := <-c.read:
		if len(buf) < len(recv) {
			fmt.Println("overflow error", len(buf), len(recv))
		}
		return copy(buf, recv), nil
	}
}
```

That's kinda silly - the goal of this package is to give messages
in byte-slice form to end users, but the current interface design
forced it to copy an already-pulled message into this second,
unneeded buffer passed into `Read` here.

So, to mitigate (somewhat) the additional memory required by the
buffer size change, this change also moves around some of the
abstractions used in the IPC package to allow for the `Read`
function above to instead just return the byte slice directly,
removing the need for a datagram server's spawned clients to
allocate their own permanent buffer. In practice, this saves us 3
buffers' worth (now 96k) of memory.

The general overview:
- Previously, `newBaseConn` would translate a `net.Conn` into our
  `Conn` interface, and the `packetClient` connections spawned by
  clients connecting to datagram servers fulfilled the `net.Conn`
  interface
- Now, `packetClient` has been modified to satisfy our `Conn`
  interface instead of being a `net.Conn`, and those clients no
  longer go through the code in `newBaseConn`
- Calls to `newBaseServer` now expect the listener passed in to
  generate `Conn` connections instead of `net.Conn`s
- The changes in `stream.go` make stream socket support a
  transparent wrapper over the underlying `net.Conn`, making it
  easier for `listenerWrapper` to use