VIC-6814 Face blank when restarting process after fault code (#3583)
The only difference between starting the processes normally vs after a fault code is that the lcd is in different sleep modes. When a fault code is displayed it is left out of sleep mode so that the code continues to be displayed. It turns out there are different reset timings depending on what sleep mode the lcd is in when reset. If in sleep mode then we have to wait 5ms before sending commands, if not in sleep mode then we need to wait 120ms before sending commands. I couldn't get reading lcd registers to work so we just always wait 120ms after reseting.