VIC-6616 Fault code DAS messages (#3536)
* VIC-6616 Fault code DAS messages

We were displaying a fault code on shutdown before the destruction of Robot. This had the potential to cause a loss of DAS messages sent during destruction as we were stopping vic-dasmgr as soon as the fault code was sent to be displayed and then calling robot destructor.

This moves the shutdown fault code displaying to happen after Robot is destroyed.

* Forces vic-dasmgr to process log messages up until the time it is signalled to stop using a special terminate log message ("@@"). This prevents us from potentially losing messages due to them not yet being processed when vic-dasmgr is told to stop.

(cherry picked from commit 2c2721475e859966827d94cd56445058c8b51d0b)
Signed-off-by: David Woods <dwoods@anki.com>