VIC-6842 Fix Text to Speech audio stream trailing silence
Currently the tts system adds the Pause Punctuation silence to the
end of the generated audio stream. The audio system can only report
tts has completed when the whole stream has been played, including
the silence. Therefore, we need to add an Acapela tag to the end of
the string we feed into the generator.

- Clean up text string
- Add silence tag to end string, silence value is set to 10 milliseconds
- Add ‘.’ to the end of strings if it doesn’t have any punctuation