VIC-6768 OTA from rc-userdev bucket instead of userdev
For the release candidate build, we should only be trying to
get delta OTA files from the `rc-userdev` bucket, not the plain
`userdev` bucket which contains `master` builds.

At the time of this commit, we are NOT building a `userdev` flavor
of release candidate.  However, it is likely that we will be.