VIC-6773 Do not pull cloud jdoc on startup (#3586)
* VIC-6773 Do not pull cloud jdoc on startup

The basic issue here is that when QA or dev runs the robot,
then OTAs/erases user data and wants to start over with a
"new life", the jdocs are still in the cloud.  So prior to this change,
on startup, the robot would see those jdocs in the cloud and
download them, overwriting the [freshly-created] ones on the
robot.  This changes that behavior so that if a newer version
is found in the cloud, we submit the robot version, thus over-
writing the version in the cloud.

Note1:  This will have to change later, when other clients besides
the robot (e.g. Customer Care) are able to change jdocs.
Note2:  This also implies that the jdoc's 'doc_version' will never
be reset to 1, since the cloud version is never really deleted.

- JdocsManager::HandleReadResponse (called soon after
  startup, when vic-cloud responds with the latest versions of the
  4 jdocs):  If the cloud version is newer, just submit the robot
  version to the cloud (using the new version number), thus
  overwriting the version in the cloud.

* VIC-6773 Treat first-time differentlly

This implements a way of knowing whether the robot has
accessed the cloud (for jdocs) since user data has been
wiped.  Thus we can distinguish between a true reset, and
another client that has changed a jdoc.

- New per-jdoc flag "has accessed cloud", initialized to false,
  serialized to disk in the jdoc file (note that since this is a new
  field in the jdoc file, and it defaults to false, then the first time
  an existing install gets this software, it will cause the jdoc to
  be written to disk at startup, but that's only one time.)
- In HandleReadResponse, soon after startup, after getting
  latest versions of all jdocs, and a jdoc has a newer version
  in the cloud, use the flag to decide whether to accept the
  cloud version (if true), or submit the disk version to cloud (if
  false).
- Also simplify and clarify the log messages
- In either case, set the flag to true, and if it was previously
  false, make sure we write the jdoc to disk to save the flag