VIC-5674 Add ability to pre-append mic data ahead of stream start time
We don’t want the earcon sound in the wake wordless stream
therefore we don’t start the stream until the earcon sound has
completed playing. However, to improve recognition we do want to
start the stream slightly earlier then when we get the earcon
completed callback. Thus we added the ability to add pre stream start
mic data to the beginning of the stream.

- Move wake wordless stream jobs to micDataProcessor class
- Consolidate wake word and wake wordless stream jobs
- Add constant for wake wordless (trigger less) overlap size
- Remove timestamp from TriggerWordDetected message, we are
  not using it and it causes confusion