VIC-6980 Override prod OTA url for ANKIDEV bots (RC branch) (#3616)
For ANKIDEV bots, we want to fetch our OTAs from the internal "rc"
bucket instead of the production site.  The production site is only for
production robots and production OTA images.  It is not for our internal
use only ANKIDEV dev-signed images.

Requires commit `3f75cd6ac2` from the `master` branch or a cherry pick
of it to actually use this setting.