VIC-5271: Convert logcat to syslog
In order to unify kernel, syslog and logcat logs, liblog was modified to forward
all logs to syslog. In the case of DAS and event messages the logs are still
sent to logd. This is done to preserve the infrastructure around DAS. This means
all non DAS logs are now only available through syslog. To handle this scripts
and utilities are converted to read log files directly.

In the case of victor_log.sh the silencing options are excluded as they are no
longer necessary.