VIC-7043 Optionally reboot after installing OTA
We may have to deploy an emergency OTA that should go into effect as
soon as the robot gets it.  To do that, this change to update-engine.py
supports a new manifest field called `reboot_after_install`.  If it is
non-zero, then we will reboot after successfully installing the OTA.