VIC-7148 update-engine: Strip "ud" from OTA url
Ooops!  When I first wrote this code, I only had to worry about the OS
version having a `d` suffix on the end for the Dev build.  Now, we also
have the "UserDev" build which gets a suffix of `ud`.  This change will
strip all the `u`s and `d`s off the right hand side of the OS version
when constructing the auto update URL.