VIC-7108 Fix robot leaving charger immediately (#3663)
If you tell vector to go back his charger, it will leave again
immediately whenever the HLAI state was previously Observing.
The transition from Observing to ObservingOnChargerRecentlyPlaced
only occurs if an OnCharger condition BecameTrueThisTick. This
BecameTrueThisTick condition is needed instead of a plain
OnCharger condition for the exploring voice command, so the robot
actually leaves the charger.
A fix went in last week to properly reset the BecameTrueThisTick,
which I guess is what was keeping the robot on the charger in other
cases.
The fix separates ExploringVoiceCommand from several other states,
so that it uses the BecameTrueThisTick{OnCharger} and others simply
use OnCharger.

(cherry picked from commit fe89845a32650eff45fa17260243a67e06df2a00)
Signed-off-by: David Woods <dwoods@anki.com>