VIC-7187 update-engine: Accept X.Y.Z.<build-num>
We want to change to 4 number versions like 1.0.0.1995.  This change
updates the regular expression that validates version numbers to accept
a 3 or 4 number version.

Later, once we are no longer producing builds with 3 number versions and
we are not downgrading to them, we can tighten up this check to only
accept 4 number versions.

(cherry picked from commit 3669ce0188a38fa72dc008ea85dc4e4daa45843e)
Signed-off-by: Stuart Eichert <seichert@anki.com>