VIC-7187 Check deploy version against Robot OS version
This cleans up the deploy.sh version check to compare the current version of code
to be deployed against the current OS version on the robot. Note that we only compare
the part of the version string that does not include build numbers. The contract going
forward is that if we break compatibility between OS / victor, we should change the version
number in some way.  Most likely this will be a change to the `patch` version, since the
major and minor version numbers are meaningful to production.

(cherry picked from commit 75546422ed420b8cb956779cebfdf812af0091a5)
Signed-off-by: David Woods <dwoods@anki.com>