VIC-7106 Reduce Cancel Timer Composite Image Message Traffic (#3666)
The CancelTimer behavior, an instance of behaviorAdvanceClock, was
sending a huge number of UpdateCompositeImage messages in a single
Engine Tick to cover every frame of the countdown animation. These
changes send the messages in batches over a number of Engine ticks to
reduce the associated message traffic and avoid swamping the
LocalUDPServer and causing 915/917 crashes.