VIC-7241 Add alternative sensory recognizer model to project
We want to do some testing to compare our 2 sensory recognizer
modes on the robot, I’ve added the second model as an option in the
mic data console var. Additionally I’ve also added the ability to change
the sensitivity of the model. To change the model and sensitivity you
go to the mic data console var section, select the recognizer model
and recognizer sensitivity enum from the drop down list then press
“UpdateRecognizerModel” button to update the recognizer system.
If you get a fail response likely the sensitivity model doesn’t exist for
the selected model, see logs for details
`MicDataProcessor.UpdateTriggerForLocale`. If you set the
sensitivity enum to “default” it will use the default sensitivity
`defaultSearchFileIndex` we have defined in
`resources/config/micData/micTriggerConfig.json`

NOTE: To make it simple to add the additional model to the project it
has been defined in the metadata as the “en-ZW” locale model.

- Update micTriggerConfig.json file to include second model
- Improve Recognizer Console Vars to include sensitivity
- Add Console Function setup to MicDataProcessor.cpp
- Add return bool to indicate the new model was set successfully
- Fix Dev Force Mic Processing state bug when going back to
  normal mic processing operation

# Conflicts:
#	animProcess/src/cozmoAnim/micData/micDataProcessor.cpp