SAIP-1694 Allow token provisioning / account creation in multiple environments
Currently the environment for token provisioning and the accounts
package is limited to "development" only.

The environment for "tokprovision" can be provided on the command line
with the "-env" flag (which defaults to "dev").

This also adds an (optional) step to the integration test suite
that creates a new test account before logging in.