SAIP-1848 Use Redis for assigning unique test users per instance
A Redis service is added to the AWS Fargate cluster that is used to
assign unique test users for every load test container instance based
on the atomic "INCR" counter.

This is used to ensure that every container is started with unique
credentials (e.g. account username and client certificates).

This also launches an admin EC2 instance for testing purposes (e.g.
to use the Redis cli).