VIC-7232 Fix windy weather anim process crash (#3756)
The `Windy` animation for weather behaviors exposed a corner case in
some NEON processor intrinsic code optimizing noise rendering on
procedural eyes. Whenever the rightmost pixel of the eye image was
less fewer pixels from the left edge of the screen than the size of a
vectorized batch processing batch, a for-loop termination condition
would fail due to an intrinsic cast caused underflow and we would
eventually seg-fault. The fix is simply to change the batch size
parameter type from unsigned to signed.

Also fixed a fallthrough bug that was resulting in errant error
prints in StreamingAnimationModifier::ApplyMessageToStreamer.