VIC-6717 Adding cropping to face detection in meet victor (#3728)
The goal of this pull request is to increase the performance of face detection during meet vector. This was tested and proved to increase the likelihood that meet vector succeeds as part of the on going work to address the issues folks are experiencing with meet vector captured in VIC-5983.

The technical aspects are the following:

 * Cropping occurs in vision system
 * Addition of a vision mode "parameter" restricts the cropping to the enroll face behavior
 * There is a slight refactor so that we can share the same face detection code in vision system
    between cropping and not cropping, I'm not married to this but this seemed a reasonable
    reduction in code duplication
 * The correction of the cropping in a separate method that "populates" the detections for others to
    consume

(cherry picked from commit 2f6e43597d4628dfd5f62c475b45e6ce268dfcea)
Signed-off-by: David Woods <dwoods@anki.com>