VIC-7085 Power down camera in power save mode (#3769)
Disabling vision component image capture will now "pause" the camera. This is done by toggling gpio pin 94 which is tied to the camera's PWDN pin. Setting it high puts the camera in power down/standby mode in which is does not capture any frames and draws considerably less power, ends up being ~40mA. The main issue with doing this is we don't know where in the capture cycle the camera is when it is put into standby. This means the first image we get after leaving standby mode is invalid and needs to be skipped. The other affect of standby mode is that camera settings are completely ignored so trying to set exposure while the camera is paused does not work. We have to bring the camera out of standby to set the exposure and then wait some amount of time for the setting to take effect (3 engine ticks seems to work) and then put the camera back into standby.

I added support to both completely enable/disable image capture as well as the ability to capture single frames at a time. This should be useful when we want to try and do occasional image processing while in power save mode.

I also added the ability for engine to request camera frames that are at or before a certain timestamp. This means engine can ask for the frame that is closest to the timestamp of the most recent state message. Sometimes we would get too new of a frame and end up dropping it so this eliminates that problem.

Also gpio.c/h were copied from robot/core so they can be ignored