VIC-6840 Report changes to eng's use of animations (#3621)
"As an animator, I want an accurate, regularly updated list of
what animations are used by engineers."

This script makes use of:
+ An existing script to find animation triggers of the form
  AnimationTrigger::* in any C++ code
+ A search for cpp/h/json files that make direct reference to
  a string like "anim_*"
+ An existing unit test to list animation triggers that
  are loaded from strings during runtime
The latter unit test requires a special build in which generated CLAD
is modified prior to compilation. This is therefore intended
to run as a nightly build.

This script outputs:
+ A list of animations referenced by name (e.g., "anim_*")
+ A list of used animation groups (e.g., "ag_*")
+ A list of AnimationTriggers used
+ A list of animation groups in the trigger map that are unused
+ A list of animation triggers in the trigger map that are unused
Assuming that the team city artifact from the last successful build
is available, it also outputs:
+ A list of animation triggers that were added, which might indicate
  usage of an existing animation in a new feature
+ A list of animation triggers that were removed
+ A list of animation groups that are newly used
+ A list of animation groups that are no longer used
+ A list of animations that are newly called by name
+ A list of animations that no longer called by name

All output is dumped to an archive that will be
available as a build artifact.