VIC-7828 face detection recognition test harness (#3826)
This is an implementation of the face tracker for testing purposes. Specifically to test the behavior enroll face, and help clarify if some of the failures to enroll faces are performance related. The performance related testing is simulated by several knobs to delay the how "fast" the detector and enrollment completes. These knobs are controlled via console variables.

This implementation is disabled by default, thus should not be reachable unless we change our build configuration.