VIC-6873: Skip Recognition Iff Face is Named (#3848)
Background: The earlier change for this ticket avoids re-running recognition once a tracked face is already recognized, i.e. once it has any data in the recognizer. But if the recognizer doesn’t immediately realize who someone is and instead _adds_ them as a session-only face, that counts as “recognized” and no further attempts will be made to see if that is actually someone _named_. That happens pretty frequently depending on when the face is first detected. So this PR changes the notion of "already recognized" for the purpose of skipping recognition to be "has a name".

* Add ability to query whether a trackingID is associated with a _named_ face in the recognizer
* Switch to treating only _named_ faces as "known" for the purposes of skipping recognition
* Update comments