VIC-7976 Fix animation streamer when looping an animation
The New Animation Callbacks were not being called when looping an
animation. We were only calling the callbacks when we changed the
streaming animation. The callbacks are used to reset state for the
next animation. The work to perform callback was moved to when the
streaming animation is set to NULL for procedural animations and in
the InitStreamingAnimation() method to assure it is called every time
an animation is played.