VIC-7927 View sprite sequences in RGB on robot (#3842)
Updating PlayAnimationByName() to accept an optional boolean input argument that indicates if the sprite sequence images should be treated as grayscale and rendered in the robot's eye hue (the default behavior). We can set that to 'false' when we want the animation streamer to render those sprite sequence images as RGB instead.