SAIP-1890 Adding voice to robot simulation and integration test
This is integrating voice functionality (i.e. connection check) into
the testable robot abstraction.

It is exposed as both a periodic action in the robot simulation as
well as a separate test case of the integration test suite.

It also adjusts the certificate file prefix format to zero-padded
with length 8 (to correspond to the new load test cert file names).