[Voice Commands] Remove “look for face” for specific behaviors (#3877)
* [Voice Commands] Remove “look for face” for specific behaviors

[VIC-6887] Remove "look for face" behavior from select voice commands

Some behaviors appeared to be not reacting quickly enough since they were trying to find a face.  We’ve decidied to remove the “find face” behavior in these instances

+ removed from BehaviorReactToHello via json
+ removed from BehaviorReactToGoodBye via json
+ removed from “show clock” intent vai BehaviorWallTimeCoordinator update

* [Voice Commands] Remove "look for face" behavior from select voice commands

[VIC-6887] Remove "look for face" behavior from select voice commands
Weather TTS and animation will now kick in regardless of the state of “find face” behavior.  We still activate the face behavior to keep us busy during TTS generation, but once the TTS is done we go right into displaying the weather

* removed some PR demo code