SAIP-1892 Add remote control to robot simulation container instances
As it takes a while for a larger fleet of instances to be provisioned
this helps to synchronize the starting / stopping of the robot
simulation so that load can be provided with more control.

It uses the Redis pub/sub mechanism to issue commands to all containers
in the load test Fargate cluster.

Also adds a docker compose file for local testing.