SAIP-1893 Add some randomization to periodic action triggering
It is now possible to express the periodic behaviour in terms of
a mean interval duration and a standard deviation. This means that
the actual period duration follows a Gaussian distribution.

This commit also allows real-time updating of the mean and standard
deviation of the periodic timers via the Redis pubsub mechanism in
order to have more control of an ongoing test.